ReasoningAgent/requirements.txt
google-generativeai
pandas
paho-mqtt